fix: avoid deleting users who were created but had errors in their sign in flow

+++ functions/src/token.ts
@@ -6,20 +6,30 @@ export const token = functions.https.onCall(async (data: { state: string }) => {
     throw new functions.https.HttpsError('invalid-argument', 'Invalid state');
   }
 
+  let stateSnap;
   try {
-    const stateSnap = await admin.firestore().doc(`/users/${data.state}`).get();
-    if (!stateSnap.exists) {
-      throw new functions.https.HttpsError('invalid-argument', 'Invalid state');
-    } else {
-      const stateData = stateSnap.data();
-      if (stateData?.error) {
-        const error = stateData?.error;
-        await admin.firestore().doc(`/users/${data.state}`).delete();
-        throw new functions.https.HttpsError('aborted', error);
-      }
-    }
+    stateSnap = await admin.firestore().doc(`/users/${data.state}`).get();
   } catch (error) {
     throw new functions.https.HttpsError('internal', error.message);
+  }
+
+  if (!stateSnap.exists) {
+    throw new functions.https.HttpsError('invalid-argument', 'Invalid state');
+  } else {
+    try {
+      await admin.auth().getUser(data.state);
+    } catch (error) {
+      if (error.code === 'auth/user-not-found') {
+         const stateData = stateSnap.data();
+         if (stateData?.error) {
+           const error = stateData?.error;
+           await admin.firestore().doc(`/users/${data.state}`).delete();
+           throw new functions.https.HttpsError('aborted', error);
+         }
+      } else {
+        throw new functions.https.HttpsError('internal', error.message);
+      }
+    }
   }
 
   let token;
